+++ src/main/java/br/com/pi/projeto_RD/repository/SubCategoriaRepository.java
@@ -1,7 +1,13 @@
 package br.com.pi.projeto_RD.repository;
 
+import br.com.pi.projeto_RD.model.entity.ProdutoEntity;
 import br.com.pi.projeto_RD.model.entity.SubCategoriaEntity;
+import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SubCategoriaRepository extends JpaRepository<SubCategoriaEntity, Long> {
+import java.math.BigInteger;
+import java.util.List;
+
+public interface SubCategoriaRepository extends JpaRepository<SubCategoriaEntity, BigInteger> {
+
 }
