+++ src/main/java/br/com/pi/projeto_RD/repository/TipoProdutoRepository.java
@@ -4,6 +4,8 @@ import br.com.pi.projeto_RD.model.entity.StatusProdutoEntity;
 import br.com.pi.projeto_RD.model.entity.TipoProdutoEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TipoProdutoRepository extends JpaRepository<TipoProdutoEntity, Long> {
+import java.math.BigInteger;
+
+public interface TipoProdutoRepository extends JpaRepository<TipoProdutoEntity, BigInteger> {
 
 }
